+++ openpecha/formatters/google_ocr.py
@@ -17,12 +17,17 @@ from openpecha.core.annotation import Page, Span
 from openpecha.core.annotations import BaseAnnotation
 from openpecha.core.layer import Layer, LayerEnum
 from openpecha.core.ids import get_base_id
-from openpecha.core.metadata import InitialPechaMetadata, InitialCreationType
+from openpecha.core.metadata import InitialPechaMetadata, InitialCreationType, LicenseType, CopyrightStatus, Copyright
 from openpecha.formatters import BaseFormatter
 from openpecha.utils import dump_yaml, gzip_str
 
 
-extended_LayerEnum = [(l.name, l.value) for l in LayerEnum] + [("low_conf_box", "LowConfBox")]
+BDR = Namespace("http://purl.bdrc.io/resource/")
+BDO = Namespace("http://purl.bdrc.io/ontology/core/")
+BDA = Namespace("http://purl.bdrc.io/admindata/")
+ADM = Namespace("http://purl.bdrc.io/ontology/admin/")
+
+extended_LayerEnum = [(l.name, l.value) for l in LayerEnum] + [("low_conf_box", "LowConfBox")] + [("language", "Language")]
 LayerEnum = Enum("LayerEnum", extended_LayerEnum)
 
 class ExtentedLayer(Layer):
@@ -30,6 +35,9 @@ class ExtentedLayer(Layer):
 
 class LowConfBox(BaseAnnotation):
     confidence: str
+
+class Language(BaseAnnotation):
+    language_code: str
 
 class GoogleOCRFormatter(BaseFormatter):
     """
@@ -43,6 +51,8 @@ class GoogleOCRFormatter(BaseFormatter):
         self.base_text = []
         self.low_conf_ann_text = ""
         self.base_meta = {}
+        self.cur_base_word_confidences = []
+        self.word_confidences = []
 
     def text_preprocess(self, text):
 
@@ -106,6 +116,26 @@ class GoogleOCRFormatter(BaseFormatter):
             return bounding_poly.get("text", "")
         else:
             return f"§{bounding_poly.get('text', '')}Ç{bounding_poly['confidence']}Ç§"
+    
+    def get_language_code(self, bounding_poly):
+        language_codes = []
+        properties = bounding_poly.get("property", {})
+        if properties:
+            languages = properties.get("detectedLanguages", [])
+            if languages:
+                for language in languages:
+                    if language_code := language.get("languageCode", ""):
+                        language_codes.append(language_code)
+        return language_codes
+    
+    def get_language_code_ann(self, bounding_poly):
+        text = bounding_poly.get("text", "")
+        language_codes = self.get_language_code(bounding_poly)
+        if language_codes:
+            language_code = "-".join(language_codes)
+            return f"§{text}Ç{language_code}Ç§"
+        else:
+            return text
 
     def get_lines(self, bounding_polys):
         """Return list of lines in page using bounding polys of page
@@ -116,27 +146,42 @@ class GoogleOCRFormatter(BaseFormatter):
         Returns:
             list: list of lines in page
         """
+        lines = {
+            'base_lines': [],
+            'low_conf_annotated_lines': [],
+            'language_code_annotated_lines': []
+        }
         prev_bounding_poly = bounding_polys[0]
-        lines = []
-        lines_with_ann = []
-        cur_line = ''
-        cur_line_with_ann = ''
+        base_lines = []
+        low_conf_annotated_lines = []
+        language_code_annotated_lines = []
+        cur_base_line = ''
+        cur_low_conf_annotated_line = ''
+        cur_language_code_annotated_line = ''
         avg_line_height = self.get_avg_bounding_poly_height(bounding_polys)
         for bounding_poly in bounding_polys:
             if self.is_in_cur_line(prev_bounding_poly, bounding_poly, avg_line_height):
-                cur_line += bounding_poly.get("text", "")
-                cur_line_with_ann += self.get_low_confidence_ann(bounding_poly)
+                cur_base_line += bounding_poly.get("text", "")
+                cur_low_conf_annotated_line += self.get_low_confidence_ann(bounding_poly)
+                cur_language_code_annotated_line += self.get_language_code_ann(bounding_poly)
             else:
-                lines.append(cur_line)
-                lines_with_ann.append(cur_line_with_ann)
-                cur_line = bounding_poly.get("text", "")
-                cur_line_with_ann = self.get_low_confidence_ann(bounding_poly)
+                base_lines.append(cur_base_line)
+                low_conf_annotated_lines.append(cur_low_conf_annotated_line)
+                language_code_annotated_lines.append(cur_language_code_annotated_line)
+                cur_base_line = bounding_poly.get("text", "")
+                cur_low_conf_annotated_line = self.get_low_confidence_ann(bounding_poly)
+                cur_language_code_annotated_line = self.get_language_code_ann(bounding_poly)
             prev_bounding_poly = bounding_poly
-        if cur_line:
-            lines.append(cur_line)
-        if cur_line_with_ann:
-            lines_with_ann.append(cur_line_with_ann)
-        return lines, lines_with_ann
+        if cur_base_line:
+            base_lines.append(cur_base_line)
+        if cur_low_conf_annotated_line:
+            low_conf_annotated_lines.append(cur_low_conf_annotated_line)
+        if cur_language_code_annotated_line:
+            language_code_annotated_lines.append(cur_language_code_annotated_line)
+        lines["base_lines"] = base_lines
+        lines["low_conf_annotated_lines"] = low_conf_annotated_lines
+        lines['language_code_annotated_lines'] = language_code_annotated_lines
+        return lines
 
     def transfer_space(self, base_with_space, base_without_space):
         """transfer space from base with space to without space
@@ -249,6 +294,11 @@ class GoogleOCRFormatter(BaseFormatter):
                         bounding_polys.append(word)
         return bounding_polys
 
+    def populate_confidence(self, bounding_polys):
+        for bounding_poly in bounding_polys:
+            self.word_confidences.append(float(bounding_poly['confidence']))
+            self.cur_base_word_confidences.append(float(bounding_poly['confidence']))
+
     def post_process_page(self, page):
         """parse page response to generate page content by reordering the bounding polys
 
@@ -258,6 +308,11 @@ class GoogleOCRFormatter(BaseFormatter):
         Returns:
             str: page content
         """
+        post_processed_pages = {
+            'base_page': '',
+            'low_conf_annotated_page': '',
+            'language_code_annotated_page': ''
+        }
         postprocessed_page_content = ""
         try:
             page_content = page["textAnnotations"][0]["description"]
@@ -265,16 +320,22 @@ class GoogleOCRFormatter(BaseFormatter):
             print("Page empty!!")
             return postprocessed_page_content
         bounding_polys = self.get_char_base_bounding_polys(page)
+        self.populate_confidence(bounding_polys)
         sorted_bounding_polys = self.sort_bounding_polys(bounding_polys)
-        lines, lines_with_low_conf_ann = self.get_lines(sorted_bounding_polys)
-        page_content_without_space = "\n".join(lines)
-        page_with_low_conf_ann = "\n".join(lines_with_low_conf_ann)
+        lines = self.get_lines(sorted_bounding_polys)
+        page_content_without_space = "\n".join(lines.get("base_lines", []))
+        page_with_low_conf_ann = "\n".join(lines.get("low_conf_annotated_lines", []))
+        page_with_language_code_ann = "\n".join(lines.get("language_code_annotated_lines", []))
         postprocessed_page_content = self.transfer_space(
             page_content, page_content_without_space
         )
         postprocessed_pg_with_low_conf_ann = self.transfer_space(postprocessed_page_content, page_with_low_conf_ann)
+        postprocessed_pg_with_language_code_ann = self.transfer_space(postprocessed_page_content, page_with_language_code_ann)
+        post_processed_pages["base_page"] = postprocessed_page_content + "\n"
+        post_processed_pages["low_conf_annotated_page"] = postprocessed_pg_with_low_conf_ann + "\n"
+        post_processed_pages["language_code_annotated_page"] = postprocessed_pg_with_language_code_ann + "\n"
+        return post_processed_pages
 
-        return postprocessed_page_content + "\n", postprocessed_pg_with_low_conf_ann + "\n"
 
     def get_input(self, input_path):
         """
@@ -296,9 +357,8 @@ class GoogleOCRFormatter(BaseFormatter):
     def extract_confidence(self, chunk):
         confidence = re.search("Ç(.+?)Ç", chunk).group(1)
         return confidence
-
         
-    def low_confidence_text_layer(self, low_conf_ann_text):
+    def format_low_confidence_box_layer(self, low_conf_ann_text):
         base_text = ""
         anns = {}
         low_conf_ann_text = low_conf_ann_text.replace("\n", "¢")
@@ -317,10 +377,128 @@ class GoogleOCRFormatter(BaseFormatter):
                 base_text += chunk
         layer = ExtentedLayer(annotation_type=LayerEnum.low_conf_box, annotations=anns)
         return json.loads(layer.json(exclude_none=True))
+    
+    def extract_language_code(self, chunk):
+        language_code = re.search("Ç(.+?)Ç", chunk).group(1)
+        return language_code
+    
+    def extract_text(self, chunk):
+        text = re.search('§(.+?)Ç.+?Ç', chunk).group(1)
+        return text
+
+    def process_first_chunk(self, chunk):
+        try:
+            cur_annotated_chunk = self.extract_text(chunk)
+        except:
+            cur_annotated_chunk = ""
+        try:
+            cur_language_code = self.extract_language_code(chunk)
+        except:
+            cur_language_code = ""
+        return cur_annotated_chunk, cur_language_code
+
+    def rm_short_ann(self, text):
+        new_text = text
+        for ann in re.findall("(§.+?§)", text):
+            ann_text = self.extract_text(ann)
+            ann  = re.escape(ann)
+            if len(ann_text) < 30:
+                new_text = re.sub(ann, ann_text, new_text, 1)
+        return new_text
+
+    def add_default_lang_code(self, text):
+        new_text = ""
+        chunks = re.split("(§.+?§)", text)
+        for chunk in chunks:
+            if chunk:
+                if re.search("(§.+?§)", chunk):
+                    new_text += chunk
+                else:
+                    new_text += f"§{chunk}ÇboÇ§"
+        return new_text
+
+    def is_mergeable_chunk(self, prev_chunk, chunk, last_lang_code):
+        if re.search("§.+?§", prev_chunk) or prev_chunk == "¢":
+            try:
+                prev_lang_code = self.extract_language_code(prev_chunk)
+            except:
+                prev_lang_code =last_lang_code
+            cur_lang_code = self.extract_language_code(chunk)
+            if prev_lang_code == cur_lang_code:
+                return True
+        return False
+
+    def merge_consecutive_ann(self, ann_text):
+        new_text = ""
+        chunks = re.split("(§.+?§)", ann_text)
+        if chunks:
+            cur_annotated_chunk, cur_language_code = self.process_first_chunk(chunks[0])
+            if not cur_annotated_chunk:
+                new_text += chunks[0]
+            prev_chunk = chunks[0]
+            for chunk in chunks[1:]:
+                if chunk:
+                    if re.search("§.+?§", chunk):
+                        if self.is_mergeable_chunk(prev_chunk, chunk, cur_language_code):
+                            if cur_annotated_chunk:
+                                cur_annotated_chunk += self.extract_text(chunk)
+                            else:
+                                cur_annotated_chunk += f"§{self.extract_text(chunk)}"
+                            cur_language_code = self.extract_language_code(chunk)
+                        else:
+                            if cur_annotated_chunk:
+                                new_text += f"{cur_annotated_chunk}Ç{cur_language_code}Ç§"
+                                cur_annotated_chunk = f"§{self.extract_text(chunk)}"
+                                cur_language_code = self.extract_language_code(chunk)
+                            else:
+                                cur_annotated_chunk = f"§{self.extract_text(chunk)}"
+                                cur_language_code = self.extract_language_code(chunk)
+                    elif chunk == "¢":
+                        if cur_annotated_chunk:
+                            cur_annotated_chunk += chunk
+                        else:
+                            new_text += chunk
+                    else:
+                        if cur_annotated_chunk:
+                            new_text += f"{cur_annotated_chunk}Ç{cur_language_code}Ç§"
+                            cur_annotated_chunk = ""
+                            cur_language_code = ""
+                        new_text += chunk
+                    prev_chunk = chunk
+            if cur_annotated_chunk:
+                new_text += f"{cur_annotated_chunk}Ç{cur_language_code}Ç§"
+            new_text = self.rm_short_ann(new_text)
+            new_text = self.add_default_lang_code(new_text)
+            new_text = re.sub("(¢)(Ç.+?Ç§)", "\g<2>\g<1>", new_text)
+        else:
+            new_text = ann_text
+        return new_text
+
+    def format_language_layer(self, language_code_annotated_text):
+        base_text = ""
+        anns = {}
+        language_code_annotated_text = language_code_annotated_text.replace("\n", "¢")
+        language_code_annotated_text = self.merge_consecutive_ann(language_code_annotated_text)
+        chunks = re.split("(§.+?§)", language_code_annotated_text)
+        for chunk in chunks:
+            if re.search("§.+?§", chunk):
+                start = len(base_text)
+                lanuage_code = self.extract_language_code(chunk)
+                base_text += re.search("§(.+?)Ç", chunk).group(1)
+                end = len(base_text)
+                span = Span(start=start, end=end)
+                uuid = self.get_unique_id()
+                language_ann = Language(span=span, language_code=lanuage_code)
+                anns[uuid] = language_ann
+            else:
+                base_text += chunk
+        layer = ExtentedLayer(annotation_type=LayerEnum.language, annotations=anns)
+        return json.loads(layer.json(exclude_none=True))
+
 
     def format_layer(self, layers, base_id):
         anns = {}
-        for (start, end, n_pg), page_ref in zip(layers["pages"], layers["pages_ref"]):
+        for (start, end, n_pg), page_ref in zip(layers["base_pages"], layers["pages_ref"]):
             uuid = self.get_unique_id()
             span = Span(start=start, end=end)
             page = Page(span=span, imgnum=n_pg, reference=page_ref)
@@ -330,7 +508,8 @@ class GoogleOCRFormatter(BaseFormatter):
         result = {
             LayerEnum.pagination.value: json.loads(layer.json(exclude_none=True))
         }
-        result[LayerEnum.low_conf_box.value] = self.low_confidence_text_layer(layers['low_conf_ann_text'])
+        result[LayerEnum.low_conf_box.value] = self.format_low_confidence_box_layer(layers['low_conf_ann_text'])
+        result[LayerEnum.language.value] = self.format_language_layer(layers['language_code_annotated_text'])
 
         return result
 
@@ -342,18 +521,23 @@ class GoogleOCRFormatter(BaseFormatter):
         return coord
 
     def _get_page(self, response):
+        pages = {
+            'base_page': '',
+            'low_conf_annotated_page': '',
+            'language_code_annotated_page': ''
+        }
         try:
             if len(response["textAnnotations"]) != 0:
                 page = response["textAnnotations"][0]
             else:
-                return None, None
+                return pages
         except KeyError:
-            return None, None
+            return pages
 
-        text, text_with_low_conf_ann = self.post_process_page(response)
+        pages = self.post_process_page(response)
         # vertices = page['boundingPoly']['vertices']  # get text box
 
-        return text, text_with_low_conf_ann  # self._get_coord(vertices)
+        return pages # self._get_coord(vertices)
 
     def _get_lines(self, text, last_pg_end_idx, is_first_pg):
         lines = []
@@ -408,8 +592,9 @@ class GoogleOCRFormatter(BaseFormatter):
 
     def build_layers(self, responses, vol_name, base_id=None):
 
-        pages = []
+        base_pages = []
         low_conf_ann_pages = []
+        language_code_ann_pages = []
         pages_ref = []
         last_pg_end_idx = 0
         img2seq = self._get_imagelist_meta(vol_name)
@@ -420,22 +605,32 @@ class GoogleOCRFormatter(BaseFormatter):
             if not response:
                 print(f"[ERROR] Failed : {n_pg}")
                 continue
-            text, text_with_low_conf_ann = self._get_page(response)
+            pages = self._get_page(response)
+            base_page = pages['base_page']
+            low_conf_ann_page = pages['low_conf_annotated_page']
+            language_code_ann_page = pages['language_code_annotated_page']
 
             # skip empty page (can be bad image)
-            if not text:
+            if not base_page:
                 print(f"[ERROR] empty page {n_pg}")
                 continue
-            lines, last_pg_end_idx = self._get_lines(text, last_pg_end_idx, n_pg == 1)
-            pages.append((lines[0][0], lines[-1][1], n_pg))
+            lines, last_pg_end_idx = self._get_lines(base_page, last_pg_end_idx, n_pg == 1)
+            base_pages.append((lines[0][0], lines[-1][1], n_pg))
             pages_ref.append(f"{page_ref}.{img2seq[page_ref]['ext']}")
 
             # create base_text
-            self.base_text.append(text)
-            low_conf_ann_pages.append(text_with_low_conf_ann)
+            self.base_text.append(base_page)
+            low_conf_ann_pages.append(low_conf_ann_page)
+            language_code_ann_pages.append(language_code_ann_page)
         low_conf_ann_text = f"{self.page_break}".join(low_conf_ann_pages)
+        language_code_annotated_text = f"{self.page_break}".join(language_code_ann_pages)
 
-        result = {"pages": pages, "pages_ref": pages_ref, "low_conf_ann_text": low_conf_ann_text}
+        result = {
+            "base_pages": base_pages,
+            "pages_ref": pages_ref,
+            "low_conf_ann_text": low_conf_ann_text,
+            "language_code_annotated_text": language_code_annotated_text
+            }
 
         return result
 
@@ -445,6 +640,83 @@ class GoogleOCRFormatter(BaseFormatter):
 
         return base_text
 
+    def get_resource_ttl(self, work_id):
+        try:
+            ttl = requests.get(f"https://ldspdi.bdrc.io/resource/M{work_id}.ttl")
+            return ttl.text
+        except:
+            return None
+
+    def get_copyright_and_license_info(self, work_id):
+        ttl = self.get_resource_ttl(work_id)
+        if ttl:
+            g = Graph()
+            try:
+                g.parse(data=ttl, format="ttl")
+            except:
+                return {}, None
+            
+            work = f"M{work_id}"
+            status = g.value(BDR[work], BDO['copyrightStatus'])
+            copyright_status = (str(status)).split("/")[-1]
+            
+            copyright_copyrighted = Copyright(
+                status=CopyrightStatus.COPYRIGHTED,
+                notice="Copyrighted by the original author or editor",
+                info_url="https://rightsstatements.org/page/InC/1.0/?language=en",
+            )
+            
+            copyright_unknown = Copyright(
+                status=CopyrightStatus.UNKNOWN,
+                notice="Copyright Undertermined",
+                info_url="https://rightsstatements.org/page/UND/1.0/?language=en",
+            )
+            
+            copyright_public_domain = Copyright(
+                status=CopyrightStatus.PUBLIC_DOMAIN,
+                notice="Public Domain",
+                info_url="https://wiki.creativecommons.org/wiki/Public_domain",
+            )
+            
+            if copyright_status == "CopyrightUndetermined":
+                license = LicenseType.UNDER_COPYRIGHT
+                return copyright_unknown, license
+            elif copyright_status == "None":
+                license = LicenseType.CC0
+                return copyright_public_domain, license
+            elif copyright_status == "Copyrighted":
+                license = LicenseType.UNDER_COPYRIGHT
+                return copyright_copyrighted, license
+        else:
+            return {}, None
+        
+    
+
+    def get_admindata_ttl(self, work_id):
+        try:
+            ttl = requests.get(f"http://purl.bdrc.io/admindata/{work_id}.ttl")
+            return ttl.text
+        except:
+            return None
+
+    def get_restrictedInChina_and_access_info(self, work_id):
+        ttl = self.get_admindata_ttl(work_id)
+        if ttl:
+            g = Graph()
+            try:
+                g.parse(data=ttl, format="ttl")
+            except:
+                return None, None
+            
+            restrictedInChina = (g.value(BDA[work_id], ADM['restrictedInChina'])).value
+            
+            output = g.value(BDA[work_id], ADM['access'])
+            access = (str(output)).split("/")[-1]
+            
+            return restrictedInChina, f"http://purl.bdrc.io/admindata/{access}"
+        else:
+            return None, None
+            
     def get_metadata(self, work_id, pecha_id):
         import xml.etree.ElementTree as ET
 
@@ -456,6 +728,13 @@ class GoogleOCRFormatter(BaseFormatter):
         bdrc_metadata_url = query_url.format(work_id)
         r = requests.get(bdrc_metadata_url)
 
+        opf_word_confidence_median = self.get_median(self.word_confidences)
+        opf_word_confidence_mean = self.get_mean(self.word_confidences)
+        
+        restrictedInChina, access = self.get_restrictedInChina_and_access_info(work_id)
+        copyright, license = self.get_copyright_and_license_info(work_id)
+            
+
         try:
             root = ET.fromstring(r.content.decode("utf-8"))
         except Exception:
@@ -465,14 +744,18 @@ class GoogleOCRFormatter(BaseFormatter):
                 imported=datetime.datetime.now(timezone.utc),
                 last_modified=datetime.datetime.now(timezone.utc),
                 parser=None,
-                copyright=None,
-                license=None,
+                copyright=copyright,
+                license=license,
+                ocr_word_mean_confidence_index=opf_word_confidence_mean,
+                ocr_word_median_confidence_index=opf_word_confidence_median,
                 source_metadata={
                     "id": f"bdrc:{work_id}",
                     "title": "",
                     "author": "",
-                    "base": self.base_meta
-                }
+                    "restrictedInChina": restrictedInChina,
+                    "access": access
+                },
+                base=self.base_meta
             )
             return json.loads(metadata.json())
 
@@ -484,17 +767,46 @@ class GoogleOCRFormatter(BaseFormatter):
                 imported=datetime.datetime.now(timezone.utc),
                 last_modified=datetime.datetime.now(timezone.utc),
                 parser=None,
-                copyright=None,
-                license=None,
+                copyright=copyright,
+                license=license,
+                ocr_word_mean_confidence_index=opf_word_confidence_mean,
+                ocr_word_median_confidence_index=opf_word_confidence_median,
                 source_metadata={
-                    "id": f"bdr:{work_id}",
+                    "id": f"bdrc:{work_id}",
                     "title": converter.toUnicode(title_tag.text),
                     "author": converter.toUnicode(author_tag.text) if author_tag else "",
-                    "base": self.base_meta
-            },
+                    "restrictedInChina": restrictedInChina,
+                    "access": access
+                },
+                base=self.base_meta
         )
 
         return json.loads(metadata.json())
+    
+    def get_median(self, list_):
+        list_.sort()
+        number_of_items = len(list_)
+        mid_index = number_of_items // 2
+        if number_of_items % 2 == 0:
+            return (list_[mid_index-1] + list_[mid_index]) / 2
+        else:
+            return list_[mid_index]
+    
+    def get_mean(self, list_):
+        grand_sum = sum(list_)
+        mean_ = grand_sum / len(list_)
+        return mean_
+
+
+    def get_base_confidence_median(self):
+        cur_base_confidences = self.cur_base_word_confidences
+        base_confidence_median = self.get_median(cur_base_confidences)
+        return base_confidence_median
+    
+    def get_base_confidence_mean(self):
+        cur_base_confidences = self.cur_base_word_confidences
+        base_confidence_mean = self.get_mean(cur_base_confidences)
+        return base_confidence_mean
 
     def set_base_meta(self, meta_ttl, image_group_id, base_file_name):
         BDR = Namespace("http://purl.bdrc.io/resource/")
@@ -515,12 +827,19 @@ class GoogleOCRFormatter(BaseFormatter):
             total_pages = int(g.value(BDR[image_group_id], BDO["volumePagesTotal"]))
         except:
             total_pages = 0
+        base_confidence_median = self.get_base_confidence_median()
+        base_confidence_mean = self.get_base_confidence_mean()
+        self.cur_word_confidences = []
         self.base_meta[base_file_name] = {
-            "image_group_id": image_group_id,
-            "title": title,
-            "total_pages": total_pages,
+            "source_metadata": {
+                "image_group_id": image_group_id,
+                "title": title,
+                "total_pages": total_pages,
+            },
             "order": volume_number,
             "base_file": f"{base_file_name}.txt",
+            "ocr_word_median_confidence_index": base_confidence_median,
+            "ocr_word_mean_confidence_index": base_confidence_mean
         }
 
     def get_meta_ttl(self, work_id):